Finished project

+++ AdditonalFunctions.h
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <iostream>
 #include <fstream>
+#include <iomanip>
 
 #ifndef INVERSEKINZADANIE_ADDITONALFUNCTIONS_H
 #define INVERSEKINZADANIE_ADDITONALFUNCTIONS_H
@@ -9,5 +10,6 @@
 using  namespace std;
 
 void WriteJointAngles(const string &filename, double *theta1, double *theta2, long long n_points);
-void writeComputingTimes(const string &filename, double timeSequence, double timeThread, long long TimeOpenMP, int Threads);
+void writeComputingTimes(const string &filename, double timeSequence, double timeThread, double TimeOpenMP, long long Threads);
+
 #endif //INVERSEKINZADANIE_ADDITONALFUNCTIONS_H
